feat(TestAutomation) Harass Calculator tests and Search box style fix

+++ Model/TravelTime.cs
@@ -2,12 +2,12 @@
 
 public class TravelTime
 {
-    public TravelTime(int index, int value)
+    public TravelTime(int index, float value)
     {
         Index = index;
         Value = value;
     }
 
     public int Index { get; set; }
-    public int Value { get; set; }
+    public float Value { get; set; }
 }
